local executor to use the same logging as remote
